requirements.txt
aiofiles==24.1.0
alembic==1.15.2
annotated-types==0.7.0
anyio==4.9.0
asttokens==3.0.0
async-property==0.2.2
async-timeout==5.0.1
asyncpg==0.30.0
attrs==25.3.0
black==25.1.0
certifi==2025.1.31
cffi==1.17.1
cfgv==3.4.0
charset-normalizer==3.4.1
click==8.1.8
cryptography==44.0.2
decorator==5.2.1
deprecation==2.1.0
distlib==0.3.9
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
exceptiongroup==1.2.2
executing==2.2.0
fastapi==0.115.12
fastapi-cli==0.0.7
filelock==3.18.0
flake8==7.2.0
flake8-bugbear==24.12.12
graypy==2.1.0
greenlet==3.2.0
h11==0.14.0
httpcore==1.0.8
httptools==0.6.4
httpx==0.28.1
identify==2.6.10
idna==3.10
ipdb==0.13.13
ipython==8.18.1
jedi==0.19.2
Jinja2==3.1.6
jwcrypto==1.5.6
Mako==1.3.10
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mccabe==0.7.0
mdurl==0.1.2
mypy==1.15.0
mypy_extensions==1.1.0
nodeenv==1.9.1
packaging==25.0
parso==0.8.4
pathspec==0.12.1
pexpect==4.9.0
platformdirs==4.3.7
pre_commit==4.2.0
prompt_toolkit==3.0.51
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.4.8
pycodestyle==2.13.0
pycparser==2.22
pydantic==2.11.3
pydantic_core==2.33.1
pyflakes==3.3.2
Pygments==2.19.1
PyJWT==2.10.1
python-dotenv==1.1.0
python-jose==3.4.0
python-keycloak==5.5.0
python-multipart==0.0.20
PyYAML==6.0.2
redis==5.2.1
requests==2.32.3
requests-toolbelt==1.0.0
rich==14.0.0
rich-toolkit==0.14.1
rsa==4.9.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.40
stack-data==0.6.3
starlette==0.46.2
tomli==2.2.1
traitlets==5.14.3
typer==0.15.2
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
uvicorn==0.34.2
uvloop==0.21.0
virtualenv==20.30.0
watchfiles==1.0.5
wcwidth==0.2.13
websockets==15.0.1
